move card/list controls to the drawer
