+++ app/client/src/views/Feed.js
@@ -1,14 +1,11 @@
 import React from 'react'
-import { Box, Grid, ToggleButtonGroup, ToggleButton, Stack } from '@mui/material'
+import { Box, Grid, Stack } from '@mui/material'
 import { useLocation } from 'react-router-dom'
-import AppsIcon from '@mui/icons-material/Apps'
-import { isMobile } from 'react-device-detect'
-import TableRowsIcon from '@mui/icons-material/TableRows'
 import VideoCards from '../components/admin/VideoCards'
 import VideoList from '../components/admin/VideoList'
 import { VideoService } from '../services'
 import LoadingSpinner from '../components/misc/LoadingSpinner'
-import { getSetting, getSettings, setSetting } from '../common/utils'
+import { getSetting, setSetting } from '../common/utils'
 
 import Select from 'react-select'
 import SnackbarAlert from '../components/alert/SnackbarAlert'
@@ -16,10 +13,7 @@ import SnackbarAlert from '../components/alert/SnackbarAlert'
 import selectFolderTheme from '../common/reactSelectFolderTheme'
 import selectSortTheme from '../common/reactSelectSortTheme'
 
-import SliderWrapper from '../components/misc/SliderWrapper'
 import { SORT_OPTIONS } from '../common/constants'
-
-const settings = getSettings()
 
 const createSelectFolders = (folders) => {
   return folders.map((f) => ({ value: f, label: f }))
@@ -31,10 +25,7 @@ function useQuery() {
   return React.useMemo(() => new URLSearchParams(search), [search])
 }
 
-const CARD_SIZE_DEFAULT = 375
-const CARD_SIZE_MULTIPLIER = 2
-
-const Feed = ({ authenticated, searchText }) => {
+const Feed = ({ authenticated, searchText, cardSize, listStyle }) => {
   const query = useQuery()
   const category = query.get('category')
   const [videos, setVideos] = React.useState([])
@@ -42,7 +33,6 @@ const Feed = ({ authenticated, searchText }) => {
   const [filteredVideos, setFilteredVideos] = React.useState([])
   const [loading, setLoading] = React.useState(true)
   const [folders, setFolders] = React.useState(['All Videos'])
-  const [cardSize, setCardSize] = React.useState(getSetting('cardSize') || CARD_SIZE_DEFAULT)
   const [selectedFolder, setSelectedFolder] = React.useState(
     category
       ? { value: category, label: category }
@@ -52,11 +42,18 @@ const Feed = ({ authenticated, searchText }) => {
 
   const [alert, setAlert] = React.useState({ open: false })
 
-  const [listStyle, setListStyle] = React.useState(settings?.listStyle || 'card')
+  const [prevCardSize, setPrevCardSize] = React.useState(cardSize)
+  const [prevListStyle, setPrevListStyle] = React.useState(listStyle)
 
   if (searchText !== search) {
     setSearch(searchText)
     setFilteredVideos(videos.filter((v) => v.info.title.search(new RegExp(searchText, 'i')) >= 0))
+  }
+  if (cardSize !== prevCardSize) {
+    setPrevCardSize(cardSize)
+  }
+  if (listStyle !== prevListStyle) {
+    setPrevListStyle(listStyle)
   }
 
   function fetchVideos() {
@@ -94,14 +91,6 @@ const Feed = ({ authenticated, searchText }) => {
     // eslint-disable-next-line
   }, [selectedSort])
 
-  const handleListStyleChange = (e, style) => {
-    if (style !== null) {
-      setListStyle(style)
-      setSetting('listStyle', style)
-      fetchVideos()
-    }
-  }
-
   const handleFolderSelection = (folder) => {
     setSetting('folder', folder)
     setSelectedFolder(folder)
@@ -112,13 +101,6 @@ const Feed = ({ authenticated, searchText }) => {
     }
   }
 
-  const handleCardSizeChange = (e, value) => {
-    const modifier = value / 100
-    const newSize = CARD_SIZE_DEFAULT * CARD_SIZE_MULTIPLIER * modifier
-    setCardSize(newSize)
-    setSetting('cardSize', newSize)
-  }
-
   return (
     <>
       <SnackbarAlert severity={alert.type} open={alert.open} setOpen={(open) => setAlert({ ...alert, open })}>
@@ -127,37 +109,6 @@ const Feed = ({ authenticated, searchText }) => {
       <Box sx={{ height: '100%' }}>
         <Grid container item justifyContent="center" spacing={2}>
           <Grid item xs={12}>
-            <Grid container sx={{ pr: 2, pl: 2, mb: 2 }}>
-              <Grid item xs sx={{ display: { xs: 'flex', sm: 'none' } }} />
-              <Grid item xs sx={{ display: { xs: 'none', sm: 'flex' } }} />
-              {!isMobile && (
-                <Grid item sx={{ pr: 2, pt: 0.25 }}>
-                  <SliderWrapper
-                    width={100}
-                    cardSize={cardSize}
-                    defaultCardSize={CARD_SIZE_DEFAULT}
-                    cardSizeMultiplier={CARD_SIZE_MULTIPLIER}
-                    onChangeCommitted={handleCardSizeChange}
-                  />
-                </Grid>
-              )}
-              <Grid item>
-                <ToggleButtonGroup
-                  size="small"
-                  sx={{ mt: -0.5 }}
-                  value={listStyle}
-                  exclusive
-                  onChange={handleListStyleChange}
-                >
-                  <ToggleButton value="card">
-                    <AppsIcon />
-                  </ToggleButton>
-                  <ToggleButton value="list">
-                    <TableRowsIcon />
-                  </ToggleButton>
-                </ToggleButtonGroup>
-              </Grid>
-            </Grid>
             <Grid container justifyContent="center">
               {videos && videos.length !== 0 && (
                 <Grid item xs={11} sm={9} md={7} lg={5} sx={{ mb: 3 }}>
